[plugin] Fix MoE profiling backend to properly detect w4a8 and w4a16 (#18)
Summary: The MoE profiler kernel checks that the weight type is uint8
and activation is bf16 for w4a16 activation and properly allocates
profiling buffers with the right sizes. This is necessary so that the
kernel accesses data similar to the real inference.

There's an issue in w4a8 quantization mode since the plugin only
provisions correct buffers if weights are the original FP4 type. This
change makes sure that we only change the type of the weight if we are
in w4a16 quant mode.

Relevant Issues: Towards GML-2616

Type of change: /kind bug

Test Plan: Tested on B200 that profiling + model works for w4a8.