[thops] Fix bug with blackwell fp8 block scale gemm (#22)
Summary: This PR fixes bugs in fp8_block_scale_gemm_blackwell_out.
* before it was erroneously throwing an exception when given a nullptr
for workspace, but if workspace size is 0 this is valid
* TrtllmGenGemmRunner::getWorkspaceSizeInBytes is required to be called
before TrtllmGenGemmRunner::run since it does more than just get
workspace size in bytes, it also configures internal state.

The biggest change this PR makes is that it changes the interface for
fp8_block_scale_gemm_out to accept torch::Tensor& for workspace instead
of a void*, which allows for communicating size information and keeps
interface types consistent. This is ABI breaking, so on the gimlet side,
when we upgrade the tllm library version, we will need to, in lockstep,
also update the callee (Fp8BlockwiseGemmPlugin).

Relevant Issues: n/a

Type of change: /kind bugfix

Test Plan: Ran plugin tests on gimlet